finished passing down props for color hash function
