+++ src/renderer/components/D3Wrapper.tsx
@@ -13,7 +13,7 @@ import ServicesWrapper from './ServicesWrapper';
 import FileSelector from './FileUpload';
 import VolumesWrapper from './VolumesWrapper';
 import ErrorDisplay from './ErrorDisplay';
-
+import { colorSchemeIndex } from '../helpers/colorSchemeHash';
 //type import
 import {
   FileUpload,
@@ -50,6 +50,16 @@ const D3Wrapper: React.FC<Props> = ({
   networks,
   uploadErrors,
 }) => {
+  const getColor: any = colorSchemeIndex();
+  const bindMountsProps = {
+    bindMounts: bindMounts,
+    getColor: getColor,
+  };
+  const dockerEngineProps = {
+    volumes: volumes,
+    getColor: getColor,
+  };
+
   return (
     <div className="d3-wrapper">
       {!fileUploaded ? (
@@ -69,8 +79,10 @@ const D3Wrapper: React.FC<Props> = ({
             options={options}
             view={view}
             networks={networks}
+            {...dockerEngineProps}
+            {...bindMountsProps}
           />
-          <VolumesWrapper volumes={volumes} bindMounts={bindMounts} />
+          <VolumesWrapper {...dockerEngineProps} {...bindMountsProps} />
         </>
       )}
     </div>
